Finished and tested the spectrum fetch and view in AM2DScanView.  The algorithm inside AMFetchSpectrumThread could and should be optimized because it goes through the file linearly.

+++ source/util/AMFetchSpectrumThread.h
@@ -27,7 +27,7 @@ public:
 
 signals:
 	/// Notifier that a new spectrum has been fetched.
-	void fetchedSpectrum(const QVector<int> &);
+	void fetchedSpectrum(const QVector<double> &);
 
 protected:
 	/// Reimplemented.  Method that is called when the thread is started.
